fix bug

+++ app/Http/Controllers/AdminController.php
@@ -31,12 +31,12 @@ class AdminController extends Controller
             'harga_beli' => 'required|numeric',
             'harga_jual' => 'required|numeric',
             'stok' => 'required|integer',
-            'gambar_barang' => 'required|image|mimes:jpeg,png,jpg,gif|max:2048',
+            'gambar_barang' => 'image|mimes:jpeg,png,jpg,gif|max:2048',
         ]);
 
-        $gambar = $request->file('gambar_barang')->store('photos', 'public');
-
+        
         if ($request->hasFile('gambar_barang')) {
+            $gambar = $request->file('gambar_barang')->store('photos', 'public');
             $imgName = 'img' . time() . '.' . $request->gambar_barang->extension();
             $request->gambar_barang->move(public_path('gambar_barang'), $imgName);
 
@@ -46,6 +46,17 @@ class AdminController extends Controller
                 'harga_jual' => $request->harga_jual,
                 'stok' => $request->stok,
                 'gambar_barang' => $imgName,
+                'activity' => $request->activity,
+                'created_at' => date('Y-m-d H:i:s')
+            ]);
+
+            return redirect()->back()->with('success', 'Data barang berhasil ditambahkan.');
+        } else {
+            Kerupuk::insert([
+                'nama_barang' => $request->nama_barang,
+                'harga_beli' => $request->harga_beli,
+                'harga_jual' => $request->harga_jual,
+                'stok' => $request->stok,
                 'activity' => $request->activity,
                 'created_at' => date('Y-m-d H:i:s')
             ]);
@@ -66,30 +77,45 @@ class AdminController extends Controller
 
         $kerupuk = Kerupuk::find($request->id);
 
-        if ($request->hasFile('gambar_barang')) {
-            $oldImagePath = public_path('gambar_barang/') . $kerupuk->gambar_barang;
-
-            $imgName = 'img' . time() . '.' . $request->gambar_barang->extension();
-            $request->gambar_barang->move(public_path('gambar_barang'), $imgName);
-
-            if (file_exists($oldImagePath)) {
-                unlink($oldImagePath);
+        
+        if($request->hasFile('gambar_barang')) {
+            if ($request->hasFile('gambar_barang')) {
+                
+                $imgName = 'img' . time() . '.' . $request->gambar_barang->extension();
+                $request->gambar_barang->move(public_path('gambar_barang'), $imgName);
+                
+                if($kerupuk->gambar_barang){
+                    $oldImagePath = public_path('gambar_barang/') . $kerupuk->gambar_barang;
+                    if (file_exists($oldImagePath)) {
+                        unlink($oldImagePath);
+                    }
+                }
+            } else {
+                $gambar = $kerupuk->gambar_barang;
             }
+            $kerupuk->update([
+                'nama_barang' => $request->nama_barang,
+                'harga_beli' => $request->harga_beli,
+                'harga_jual' => $request->harga_jual,
+                'stok' => $request->stok,
+                'gambar_barang' => $imgName,
+                'activity' => $request->activity,
+                'updated_at' => date('Y-m-d H:i:s')
+            ]);
+    
+            return redirect()->back()->with('success', 'Data barang berhasil diperbarui.');
         } else {
-            $gambar = $kerupuk->gambar_barang;
+            $kerupuk->update([
+                'nama_barang' => $request->nama_barang,
+                'harga_beli' => $request->harga_beli,
+                'harga_jual' => $request->harga_jual,
+                'stok' => $request->stok,
+                'activity' => $request->activity,
+                'updated_at' => date('Y-m-d H:i:s')
+            ]);
+
+            return redirect()->back()->with('success', 'Data barang berhasil diperbarui.');
         }
-
-        $kerupuk->update([
-            'nama_barang' => $request->nama_barang,
-            'harga_beli' => $request->harga_beli,
-            'harga_jual' => $request->harga_jual,
-            'stok' => $request->stok,
-            'gambar_barang' => $imgName,
-            'activity' => $request->activity,
-            'updated_at' => date('Y-m-d H:i:s')
-        ]);
-
-        return redirect()->back()->with('success', 'Data barang berhasil diperbarui.');
     }
